Style changes
Changed wording and size of details button on catalogue page

Linked the tables stylesheet to the myClasses page (can always remove in the future if not needed)

Adjusted group message, spacing, text alignment, and added a link to catalogie for ease of access. Also imported GroupPage stylesheet for future use.

Removed width to 90% CSS line in HomePage styling